requirements.txt
configparser
paho-mqtt